+++ Modules/Intent.Modules.Application.DomainInteractions/release-notes.md
@@ -1,5 +1,6 @@
 ### Version 1.1.3
 
+- Improvement: Support for Static Constructor mapping.
 - Fixed: Logic to use AutoMapper would be generated even when the required AutoMapper module was not installed.
 
 ### Version 1.1.2